RatingCalculator.php: made it less likely for the RatingCalculator to error
This was in response to finding some error messages on Blue Host
suggesting that events wasn't the expected data type and that division
by zero occasionally happened.

For the error messages about data type, it may be possible to look
deeper and resolve the problem with a PHP config for json_decode but I
didn't want to spend the required time for that just now.